+++ rackspace_monitoring/drivers/rackspace.py
@@ -28,7 +28,7 @@ from rackspace_monitoring.utils import to_underscore_separated
 from rackspace_monitoring.utils import value_to_bool
 
 from rackspace_monitoring.base import (MonitoringDriver, Entity,
-                                       NotificationPlan, MonitoringZone,
+                                       NotificationPlan, MonitoringZone, PrivateZone,
                                        Notification, CheckType, Alarm, Check,
                                        NotificationType, AlarmChangelog,
                                        LatestAlarmState, Agent, AgentToken,
@@ -254,7 +254,8 @@ class RackspaceMonitoringDriver(MonitoringDriver, OpenStackDriverMixin):
                             (response.status, value_dict['url'], details))
 
     def _plural_to_singular(self, name):
-        kv = {'entities': 'entity',
+        kv = {'private_monitoring_zones': 'monitoring_zone',
+              'entities': 'entity',
               'agent_tokens': 'agent_token',
               'agents': 'agent',
               'alarms': 'alarm',
@@ -406,11 +407,23 @@ class RackspaceMonitoringDriver(MonitoringDriver, OpenStackDriverMixin):
                       'list_item_mapper': self._to_datapoints}
         return LazyList(get_more=self._get_more, value_dict=value_dict)
 
+    ####################
+    ## Monitoring Zones
+    ####################
 
     def _to_monitoring_zone(self, obj, value_dict=None):
-        return MonitoringZone(id=obj['id'], label=obj['label'],
-                              country_code=obj['country_code'],
-                              source_ips=obj['source_ips'],
+
+        if ( obj['id'].startswith('mz') ):
+            return MonitoringZone(id=obj['id'], label=obj['label'],
+                                  country_code=obj['country_code'],
+                                  source_ips=obj['source_ips'],
+                                  driver=self)
+
+        return PrivateZone(id=obj['id'], label=obj['label'],
+                              maximum_checks=obj['maximum_checks'],
+                              maximum_agents=obj['maximum_agents'],
+                              disabled=obj['disabled'],
+                              metadata=obj['metadata'],
                               driver=self)
 
     def list_monitoring_zones(self):
@@ -422,6 +435,24 @@ class RackspaceMonitoringDriver(MonitoringDriver, OpenStackDriverMixin):
         url = '/monitoring_zones/%s' % (monitoring_zone_id)
         resp = self.connection.request(url).object
         return self._to_monitoring_zone(obj=resp)
+
+    def delete_monitoring_zone(self, monitoring_zone, **kwargs):
+        return self._delete(url="/monitoring_zones/%s" % (monitoring_zone.id),
+                            kwargs=kwargs)
+
+    def update_monitoring_zone(self, monitoring_zone, data, **kwargs):
+        return self._update("/monitoring_zones/%s" % (monitoring_zone.id),
+            data=data, kwargs=kwargs, coerce=self.get_monitoring_zone)
+
+    def create_monitoring_zone(self, **kwargs):
+        data = {'who': kwargs.get('who'),
+                'why': kwargs.get('why'),
+                'label': kwargs.get('label'),
+                'maximum_checks': kwargs.get('maximum_checks'),
+                'maximum_agents': kwargs.get('maximum_agents')}
+        headers = kwargs.get('headers', {})
+
+        return self._create("/monitoring_zones", data=data, coerce=self.get_monitoring_zone, headers=headers)
 
     ##########
     ## Alarms